// src/generated/api/returns/models/returnSummaryDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */
import type { ReturnStatus } from './returnStatus';
import type { ReturnChannel } from './returnChannel';
import type { ReturnReasonCode } from './returnReasonCode';
import type { ReturnSummaryDtoFault } from './returnSummaryDtoFault';

export interface ReturnSummaryDto {
  /** @pattern ^[1-9][0-9]*$ */
  id: string;
  returnNo: string;
  /** @pattern ^[1-9][0-9]*$ */
  orderId: string;
  orderNo: string;
  status: ReturnStatus;
  channel: ReturnChannel;
  reasonCode: ReturnReasonCode;
  /** @nullable */
  fault?: ReturnSummaryDtoFault;
  recipientName: string;
  itemCount: number;
  createdAt: string;
  version: string;
}